[FIX] appointment: set attendee as organizer by default
The logged user was being set as an organizer by default even though
a particular attendee's slot had been clicked on the gantt view.
so the attendee will be set as an organizer by default when clicking
on the slot.

A test case has been added to ensure the above functionality for
the gantt view does not break.

Task-3679307

closes odoo/enterprise#56123

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>