[FIX] l10n_co_edi: correct calculation of withholding tax base
Context
-------
In Colombia, a withholding tax is applied to the VAT, calculated as a
percentage of the VAT amount. A typical scenario involves a VAT at 19%
and a withholding tax at 15% of the VAT's 19%. The existing system
constraints prevent directly using the value of one tax as the base for
another, leading to a workaround by setting it to -2.85 (representing
15% of 19%).

Problem
-------
The electronic invoice requirements mandate the submission of base
amounts and taxed values for each tax and invoice line. Due to our
system's limitation in directly calculating the base for the withholding
tax, our approach has been to reverse calculate the base using the tax
amount divided by its rate. This method introduces inaccuracies because
the tax amount is rounded, and those inaccuracies can in turn result in
the electronic document being rejected.

Solution
--------
There's currently no way to properly fix this, so we have to rely on
some dodgy programming. This commit changes the calculation method to
focus on directly determining and computing the VAT amount subject to
withholding.

opw-3744872

closes odoo/enterprise#59207

X-original-commit: 1ef7118c9de82e77722254e9a885bd61ca5796fe
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Séna Serge Nshimiyimana (sesn) <sesn@odoo.com>