[FIX] account_reports: prevent reset to draft if subsequent closing entries
Prevents the user to reset to draft a tax closing entry if subsequent closing entries are already posted. This way, the user is forced to reset to draft progressively back in time the closing entries. Accounting wise, this case should not happen. It is by the way prevented by the irreversible_lock_date module (which can be however uninstalled). This mechanism solves an issue with carryovers that need to be progressively recomputed when resetting to draft a closing entry.

task-3520338

closes odoo/enterprise#59534

X-original-commit: a5afa4b21d47b1b82a4a4ce66d67d3cd6e6585f1
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>