[FIX] appointment: check the presence of attendees
This traceback will arise when the user removes the attendees while creating the
new "Staff Booking".

Steps to reproduce:

- Install "appointments"
- Open appointments -> Dental care
- Go to new -> create appointment -> remove Attendees -> Save&Close

Traceback : "TypeError
'bool' object is not subscriptable"

This commit will check for the presence of partner_ids. If they are present, it
will schedule the appointment; otherwise, it will not.

sentry-4934421570

closes odoo/enterprise#60154

X-original-commit: a10825fe3b129f8be1e4023b7364874d38a9563e
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>