[FIX] account_budget: aggregate percentage by avg
**Current behavior:**
On the pivot view for the budget analysis view, the percentage
column total is the sum of each row's percentage.

**Expected behavior:**
This value is the mean of all percentages.

**Steps to reproduce:**
1. In the Accounting app, create a budget that has a non-zero
     Achievement value in at least 2 rows

2. Go to *Reporting* -> *Budgets Analysis* and expand the y-axis
     to show the lines created, observe that the Achievement
     total is displaying the sum of all percentages

**Cause of the issue:**
The percentage field is being aggregated by summation in the
read_group() method of the `crossovered.budget.lines` model.

**Fix:**
Calculate the mean instead of the sum.

opw-3761952

closes odoo/enterprise#60431

X-original-commit: 33edb242b8c8b52f4761e58566b653e1118b6a9a
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>