[FIX] l10n_ma_reports: branch access rights
Since this pr: https://github.com/odoo/enterprise/pull/51954, multiples test
from test_company_branch were failing due to access rights.

The reason is that in l10n_ma_report, we used a function called
_get_reconciled_payments, that is a helper to retrieve the reconciled payments
on a journal entry. This method will call _get_reconciled_amls (to retrieve the
reconciled move lines on a journal entry) and then return the payment_id from
the move.

When being in the branch environment with an invoice from the branch and a
refund from the root company, doing the reconciliation between the two will
cause an access right error.

The way we solved this, is to add a sudo so that we can access the move without
having an error.

closes odoo/enterprise#59195

Task: 3802369
X-original-commit: 17bf4fddfc1e7164e94a324075d37a26e5f0d378
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>