[FIX] mrp_workorder: save workorder move to other workcenter in backend
Steps to reproduce the bug:
- Create a storable product P1 with BoM:
    - Workorder: OP1 workcenter: Assembly 1

- Create a MO with P1 and confirm it
- Go to the shop floor
- Select “Assembly 1” and “Drill 1”
- Go to assembly 1 > select the workorder OP1
- on the right side of the footer in as ": (option)" button, which opens
a pop-up window with additional options for the work order

- click on “Move to workcenter”
- Select “Drill 1”

Problem:
Change is applied in the frontend view (JS side) but not in the backend.
If you refresh the page, the change will be lost. This is due to this
commit:

https://github.com/odoo/odoo/commit/6bd2c1fdfb5722b1f2d4fc2684f20ed067b93b7d
The purpose of this commit is to no longer save the record of an x2many
when an autosave field is applied to an update

Solution:
Revert the changes made in this commit:
https://github.com/odoo/enterprise/commit/518dcd935094e30c8a39982be9b3ddb161bd03cd#diff-7df14ba8355f342f6f776e8ea293eb97ab75660c0f57245f5941e8d1da9a49bbR42-R45

opw-3816964

closes odoo/enterprise#59389

X-original-commit: 8270820b2088b92cd2996224a1897973e26db5a6
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Djamel Touati (otd) <otd@odoo.com>