[FIX] mail,*: race condition in posting message
Currently, posting a message to discuss channels is done in two steps:
first writing the last_interest_dt to the channel, then creating the
message, second triggering the notify_thread to send the message
to the followers.

In the first step, the last_interest_dt will be directly sent to the
client if it differs from the old value. So there is no need to send
the message to the client if the last_interest_dt has not changed in the
second step. Also, this can lead to a racing condition in the testing
files.

This commit removes the notif in the second step.

Also, adapting the mock_models to the new behavior as the follow-up of
https://github.com/odoo/odoo/pull/155569

closes odoo/enterprise#59188

X-original-commit: efe1c1c1ef3a34c9629ab82dfe6bbeaa5a06c99c
Related: odoo/odoo#158746
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>
Signed-off-by: Zelong Lin (zel) <zel@odoo.com>