[FIX] stock_barcode_mrp: use correct uom in MO with barcode
Issue:
======
Creating a MO with barcode uses the wrong quantity for the components.

Steps to reproduce the issue:
=============================
- Install mrp , barcode
- Activate unit of measure
- Create a product and assigna barcode number to it
- Create a BoM for that product with a component `comp`
- The `comp` product have `g` as unit of measure
- Assign the unit of measure in the BoM for the `comp` as `kg`
- Go to barcode -> operations -> manufacturing
- Create a new order
- Enter barcode manually of the product that you set up early
- Click on +1 button
- You can see the quantity is wrong as and not counting the difference
  of UoM

Solution:
=========
We need to fetch the uom data of the move to compare it to the one of
the product and update the quantity accordingly and use the uom of the
product in the bom line instead of the product.

opw-3776896

closes odoo/enterprise#63103

X-original-commit: 1005ab79241a023ca4b12eef395e9998cf981f0f
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Mahdi Cheikh Rouhou (macr) <macr@odoo.com>