[FIX] l10n_mx_edi_pos: fix undeterministic error in tour
Since 39a0542b6505e18e3fdfbefba4210d8ba87bae3c, the value of
`l10n_mx_edi_cfdi_to_public` depends on the partner linked to the
pos_order. For instance, a mexican partner without VAT will have
`l10n_mx_edi_cfdi_to_public` to True by default.

The tour `test_mx_pos_invoice_previous_order` is failing
undeterministically. The tour proceeds as follows:

- create an order without partner, and validate it
- later, go the list of paid orders, select the last one and invoice it
- the list of partners opens to allow setting one
- then, a dialog opens to set the `l10n_mx_edi_usage` and
`l10n_mx_edi_cfdi_to_public`

Currently, we simply select the first partner in the list ("(AR) Exento"
if l10n_ar is installed...). When selecting this partner, we trigger the
compute which will set a value on `l10n_mx_edi_cfdi_to_public`. Then, we
can input `l10n_mx_edi_usage` and `l10n_mx_edi_cfdi_to_public` in the
dialog. But the values filled in that dialog are not always saved to the
database before ending the tour.

This fix uses a new partner to no longer depends on demo data. In
addition, we leave the default values in the dialog to make sure the
test will not fail no matter if the dialog's values are saved or not in
the database (in a proper subsequent fix, we should ensure that we
always end the tour after the dialog's values are saved).

runbot build error 60671

closes odoo/enterprise#59310

X-original-commit: 2be6e64ef23717ab36dbba87723db666a1757eee
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>