[IMP] pos_iot: Delete Worldline LTS
We currently have an unnecessary and unused feature for Worldline "Last
Transaction Status". It's a button which checks the status of the last
transation processed by the terminal and sends this data to the PoS
frontend.
This button adds some unnecessary complexity to our codebase and is
widely unused by the clients, so we are removing it in the latest
version.

task-3460003

closes odoo/enterprise#59535

Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>