[FIX] industry_fsm_stock: Fix singleton error
Steps to reproduce:
In field service, have a project on which you have to sell 3 of item A. Item
a is tracked by serial number and the delivery route is in 2 steps. From stock
 to output and from output to customer.

On the sale order make sure you have 3 separate lines of 1 quantity of product
A separated or not by sections.
Confirm this one. This will create two pickings.
One picking linked to a global stock move for 3 units from stock to output.
A second picking linked to 3 stock moves each linked to a sale line id.
Prioritise the first one by starring it.

Now go back to the project and mark is as done.

OPW-3792527

closes odoo/enterprise#60083

X-original-commit: 50c58d4c482e731c4f34038585f36fb5c548e5b1
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>