[FIX] stock_barcode: fix broken links on homepage and settings
The purpose of this commit is to fix the broken links on the homepage of
the barcode app in addition to the print links in the settings of
Inventory and Manufacturing apps related to barcodes. As dictated by
odoo/enterprise#58078, urls now have to start with a "/" in order to
work properly.

Task-3821539

closes odoo/enterprise#59277

X-original-commit: 227137031632f22d7e4c27d301f51852ec192e80
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Mohammad Abdulmoneim (abdu) <abdu@odoo.com>