[FIX] documents: fix update document in activity view
Bug 1: How to reproduce:
1. Schedule an activity for a document.
2. Go to activity view.
3. Select the document to preview in the inspector.
4. Rename the document.
5. Refresh the page.

The document name is not renamed.

Similarly to what is done in document kanban and list controller where updating
the document through the inspector works, we activate the parameter multiEdit
which solve the problem.

Bug 2: How to reproduce:
1. Schedule an activity for a document.
2. Go to activity view.
3. Select the document to preview in the inspector.
4. Erase the name of a doc and click anywhere to save.

No alert pop-up is shown up while saving with an empty name while this is the
case when editing the document from the kanban or the list view.

Similarly to what is done in document kanban and list view, we add the required
attribute to the field name.

Task-3817795

closes odoo/enterprise#61792

X-original-commit: 05ec1058b50f006ce3952449b350f841edc9f585
Related: odoo/odoo#164052
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>