[FIX] documents_spreadsheet: sharing workspace with only 1 file will share workspace
Issue: when trying to share the spreadsheet workspace with only one
document on it, instead of selecting the right view for the workspace
we currently share the document view directly without giving the user
any change to get to the workspace view where we are able to upload new
file if allowed or see the rest of documents in the folder.

Steps to reproduce:

1. Install Documents and go to Spreadsheet.
2. Create or Upload a new spreadsheet.
3. Click next to the upload button and in the dropdown, select share
to open the share popup.
4. Inside here just select the allow to download and upload documents
and we click on share again.
5. Open the link in a new tab and we will see only the document already
opened, without any posibility of see the entire folder or uploading a
new document.

Solution: Inside the `share_portal` of `SpreadsheetShareRoute` when
checking for a single document, we also need to check if we are working
with a selection, or if it's the folder what we want to share.

opw-3688409

closes odoo/enterprise#60086

X-original-commit: a59b1bdb5a8663620048a06f9f8cbcd0a051cc46
Signed-off-by: Maruan Aguerdouh Mohtar (magm) <magm@odoo.com>