[FIX] stock_barcode(_mrp): back button
How to reproduce:

- Open the Barcode app;
- Click on Operations > Receipts > open any receipt;
- Refresh the webpage (press F5 or whatever);
- Click on the back button.

=> Expected behavior: We go back on the Receipts kanban view.

=> Current behavior: The form view of the selected picking is oepened.

The cause is, since odoo/odoo#142007, the breadcrumb is reconstructed
from the URL. But after a refresh there is no way to make a distinction
between actual form view and active id. So, when restoring the previous
controller state, since the real ones where lost, it uses the
reconstructed ones and the form view of the active picking is supposely
the previous controller state.

To fix state, in such case we call native `history.back` instead which
makes the browser to move on the previous visited page.

To be consistent, the same method is now called both when the back
button is clicked or when the `history-back` event is triggered.

Also, the old behavior of displaying Barcode main menu in case the
breadcrumb only contains 1 step is kept because it can still happen in
some case when the picking is open right away in the Barcode app by
typing the URL directly (like it's done in the tour for example.)

This issue can also apply when an operation is validated or canceled.
With this commit, the same method will be called in any case so the
behavior will stay consistent.

X-original-commit: 6772907a14d8a3b93403676a6c5a6b89ea91a0cb
Part-of: odoo/enterprise#59805