[FIX] account_reports: include VAT Unit in PDF
**Current behavior:**
If a company belongs to a VAT Unit and a report is generated
with that Unit selected, the PDF output will display the company
VAT number instead of the VAT Unit number.

**Expected behavior:**
The VAT Unit number will be written to the PDF if a report is
being generated with that option selected.

**Steps to reproduce:**
1. Install l10n_be and in the company config for the default
     Belgian company, add a new VAT Unit and make the Belgian
     company part of it

2. In the Accounting app, go to the `Tax Report` view under the
     `Reporting` menu

3. Select the newly created VAT Unit for the `Tax Unit` option
     in the top right of the view

4. Select the `PDF` button to print the report, observe that the
     `VAT:` label on the report shows the Company VAT number
     instead of the Unit's

**Cause of the issue:**
The XML template for the PDF report is only setup to show the
company VAT number here.

**Fix:**
Use the selection for the `Tax Unit:` configuration to inform
the VAT number that is shown on the final report.

opw-3747319

closes odoo/enterprise#59240

X-original-commit: 5443639767b0ff707c7074c579bd86c8b772fbeb
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Vincent Ethan <etvi@odoo.com>