[FIX] sign: traceback duplicating template
Before this commit, when duplicating a read-only template a traceback
was being triggered since we were passing a single template to the
copy method.

After this commit, we call the copy function passing a batch of a
single element, which will return a list of templates containing a
single template. After that, we call the action to open this copied
template.

task-3790206

closes odoo/enterprise#59390

X-original-commit: 758e8d64b0335167da169babc79288290f33e461
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Gabriel de Paula Felix (gdpf) <gdpf@odoo.com>