[FIX] documents_spreadsheet: inverse access check
The ir.rule in documents_spreadsheet opens too many doors.

With the condition `'|', ('document_id', '=', False)`, the rule result
is always `True` for comment threads which are not linked to a document
(but linked to a dashboard or a spreadsheet template).

Since all rules are OR-ed together, every operation is therefore allowed
on threads on dashboards and templates, regardless of the user's rights
on the underlying dashboard or template record.

closes odoo/enterprise#59253

Task: 3822402
X-original-commit: 69c85ff53cd9274bf3af5c9fd6e3f66ae3299897
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>