[REV] stock_barcode*: menu button
Undo 02edbc5a8bee2f37a6aeb2f5c88fd8e57f21007f because it useless since
odoo/odoo#142007

This button was here to always give a way to go back to main menu and
exit the Barcode app because, since the Barcode app is displayed in
fullscreen, if there is no breadcrumb (after a refresh for example),
there is no path to quit the application.
But thanks to the linked PR, now the breadcrumb is re-created from the
URL so the initial issue is obsolete.

X-original-commit: 8ffac32f64b95a796840f9704c2c2f3aae948b22
Part-of: odoo/enterprise#59805