[FIX] sale_planning: does not override planned dates in default values
Before this commit, when the user clicks on a gantt cell of the planning
gantt view to create a shift for a specific date. By doing that, the
planned dates set by default to the form view is the whole period of
the gantt view instead of the date set on the gantt cell clicked.

This commit removes the override erasing the default planned dates by
the period of the gantt to keep the standard behavior as before.

closes odoo/enterprise#60292

X-original-commit: e04b3a97cd00d5eb0c983272884546fd44089e5a
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>