[FIX] account_reports: General Ledger: add move_name to order of shown move lines
To reproduce the issue:

1) Create an account.move move_1, leave it in draft
2) Create another account.move move_2, using the same accounts as move_1 and at the same date, leave it in draft
3) Make sure move_1 and move_2's sequence numbers are empty (so that their sequence number will be recomputed at posting)
3) Post move_2
4) Post move_1
5) Open the General Ledger, unfold the lines of the accounts touched by move_1 and move_2.
===> Problem: the move_1's move lines appear before move_2's, even though move_2 has a lower sequence.

This is due to the fact move lines with identical dates were ordered by id by the GL. Since move_1 was created first, its lines had a lower id.
We fix that by adding the move_name to the ordering criteria. Move lines will still be primarily ordered by date, but if they have the same date, we'll now use the move_name before falling back to their id.

closes odoo/enterprise#60490

X-original-commit: d4b78066cb88f1d63d5411230fe478d52d3e2ed0
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>