[FIX] documents_hr: fix documents smart btn in profile
HOW TO REPRODUCE
================

Log in as an internal user who is not a member of the `documents_user` group.
Go to your profile. You should see the "Documents" smart button.

However, the button has a `groups` attribute to restrict its visibility
to the members of the `documents_user` group.

HOW TO FIX
==========

This issue is already known and a tolerated side-effect
(see commit 3651a3721865a465f5457b67436939a91f904bd1 in PR odoo/odoo#29659)

Moreover, it makes sense for any internal user to be able to access the
Documents application, whether they have rights on it or not. As an
employee, they may have documents they need to access (e.g.: contract,
payslip...)

The solution would be to remove the `groups` attribute from the smart
button to reduce confusion among developers. And also, clarify that,
indeed, any internal users can see the "Documents" smart button in their
profile.

task-3688377

closes odoo/enterprise#61227

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>