[REF] account: `odoo.tools.SQL`-ify _get_query_tax_details
In the related community PR:
The function `_get_query_tax_details` was changed
to use SQL objects for its parameters and return value
(instead of strings for queries and lists for to be injected parameters).

Here the usages of this function were adapted.

See the related community PR for more details.

task-3791228

closes odoo/enterprise#59267

Related: odoo/odoo#158880
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>