[IMP] sign: request UI improvements
This commit updates the user interface of the Request Signature form by
removing the group labels Signers and Options, putting the Attachments
button icon and label in the same line and moving the Signing Order field to
below signers emails. Additionally, in the card view of Sign Documents, the
signees color is changed from green to black, the checkbox color is changed
from black to green and the checkbox is now horizontally aligned with the
remaining fields.

task-3750770

closes odoo/enterprise#58090

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>