[FIX] account_report: rounding on no figure type
When an expression had no figure type, the condition "expression.figure_type != 'string'"
was true and so we tried to round the value. But when the value is None, the
code will crash.

Now if the expression has no figure type we get the column type of the
expression, but it can happen that an expression is not link to a column. If
that's the case we don't round the expression.

Task: 3776104
X-original-commit: d76587129e7fa6c4a2a88e37b817dc89115bc74a
Part-of: odoo/enterprise#59567