[FIX] l10n_cl_edi_website_sale: breaking checkout flow
Chilean ecommerce checkout flow was consistently breaking when
automatic invoice was enabled (Sale settings), for customers choosing
the 'Ticket' option (not the 'Invoice' one).

Indeed, choosing the 'Ticket' options sets a generic 'Anonymous' customer
as billing address, but since this address is incomplete, the customer
is redirected to the page to complete it, even though they don't have
the rights to update that address since it's not theirs.

This commit makes sure to consider this 'anonymous' billing address as
being always complete, even if some fields are empty.

opw-3792844

closes odoo/enterprise#59351

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>