[FIX] l10n_mx_edi_pos: missing await when setting MX fields
This tour was failing undeterministically because the values set in the
dialog were not saved to the database.

A naive fix was quickly merged in
2be6e64ef23717ab36dbba87723db666a1757eee: we kept the default values in
the dialog (so it was no longer failing even if the values weren't saved).

This actually occurs because we call `action_pos_order_invoice` before
saving the values of the dialog to the server. Adding an `await` fixes
this. Hence we can set a random (different from the default)
`l10n_mx_edi_usage` and `l10n_mx_edi_cfdi_to_public`.

runbot build error 60671

closes odoo/enterprise#59508

X-original-commit: 4e7a94893bda39d1951a58e08297cc8bd2296a92
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>