[FIX] *: add missing props to client actions
Since [1], it's now possible for a client action to use the resId and
update the resId on the URL as a form view will do. To do this two props
are given to the client action: resId and updateResId.

This commit adds the missing props to existing client actions that were
forgotten by the initial commit.

[1] https://github.com/odoo/odoo/commit/3ad4fd65387f60b524e5f786556963ead8ae9dfe

closes odoo/enterprise#59392

X-original-commit: 227460b703ee74a694f755f6e5e5b40a8ffba89b
Related: odoo/odoo#159168
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>