[FIX] planning: ensure timezone used is UTC in the whole test
Before this commit, the test fails since 23 March 2024 because next week
the hours changed (summer time). The problem is there if the timezone is
not UTC.

This commit ensures the whole test uses the same timezone and only UTC
one to avoid having to manage summer/winter time in the test.

runbot-60932

closes odoo/enterprise#59678

X-original-commit: e6cd3b60f6ada560a1a244ce983e05a832d593f2
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>