[FIX] sale_subscription: prevent change `recurring_invoice` value
Issue:
------
It is possible to change the recurrence of a product if it has
already been sold (correct behaviour).

The customer receives a warning, but the `recurring_invoice` field is still
set to `True`,
whereas it may have been set to `False` before the operation.

Solution:
---------
Prevent the `recurring_invoice` field from being changed if a confirmed
sale order line contains this product, to avoid confusion.

opw-3457160

closes odoo/enterprise#59367

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>