[MOV] *spreadsheet*: Move pivot code to o-spreadsheet
This commit is the enterprise counterpart of https://github.com/odoo/odoo/pull/159259.

This commit prepares the ground to the introduction of the spreadsheet
pivot table in github.com/odoo/o-spreadsheet. It's basically a move of
the plumbing necessary to make a pivot works from odoo codebase to
o-spreadsheet codebase.

A note on this commit:

- Tests are kept in Odoo

Moving tests would require to make a dummy pivot in o-spreadsheet, which is
not ideal. Let's write tests during/after the introduction of spreadsheet
pivot.

closes odoo/enterprise#59423

Task: 3748583
Related: odoo/odoo#159259
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>