[FIX] survey: support i18n for radio_selection_with_filter
The "radio selection with filter" widget allows to show/hide some
selection values given a list of allowed selection values. The feature
is similar to the "filterable selection" widget but for radio buttons.

The allowed selection context now takes a list of technical strings
instead of a list of (technical string, human string) pairs. This
allows to keep the (possibly translated) original values along with the
original selection order.

task-3829801

closes odoo/enterprise#60839

X-original-commit: b43367377914c5c9a1aadf1d919c7c2559bf7948
Related: odoo/odoo#162051
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Julien Castiaux (juc) <juc@odoo.com>