[FIX] mrp_workorder: real cost in MO overview
This commit contains small changes to adapt to the community commit.

Community commit message:
------------------------

The mo overview costs are not accurate.

Problems
--------

- as soon as the MO is in progress, the real costs should start at 0, and the
component cost should not be added untill they are consummed
- when the MO is draft/confirmed, you should not see the real cost, but the bom cost
to be able to compare it to what component are used/operations are planned on the MO
- when the MO is done, you want to compare the real cost vs what was planned

closes odoo/enterprise#67802

Task-id: 3369946
Related: odoo/odoo#175329
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>