[FIX] timer: Add model into the condition when searching timer records
When a task and a ticket share the same ID, the _search_user_timer_id method will not work and provide the wrong result when using User Timer as the filter.

Steps to reproduce:
Having a task and a ticket with the same ID.
Start the timer in the task.
Go to the ticket and filter by "User Timer is set"
There will be a ticket without a timer running but still showing in the search result.

Cause
The search method did not take into consideration of model.

Solution
Add model as a condition when searching timer records

closes odoo/enterprise#59398

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>