[FIX] pos_iot: correctly compute use_proxy in pos
Current behavior:
After this refactor https://github.com/odoo/odoo/pull/142566 the
use_proxy became undefined. And so the pos is not able to connect to the
proxy and use the printer connected to the IoT Box.

Steps to reproduce:
- Install the PoS IoT module
- Configure the IoT Box with a printer
- Select the printer from the IoT Box in the PoS configuration
- Try to print a receipt
- Nothing happens, it only try to print via the browser print popups

Note:
There was also an issue in the StatusLoop that was not correctly getting
the ids of the devices. This was fixed in the same commit.

opw-3800437

closes odoo/enterprise#59929

X-original-commit: 921e45170220c1ba3fd9c0df1e5e8d1d600b74a6
Related: odoo/odoo#160176
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Robin Engels (roen) <roen@odoo.com>