[FIX] account_report: fix xpath
During the mentioned commit (https://github.com/odoo/enterprise/pull/45792/commits/3a3d9c877f3e8bd2db9e66cd8120d889fd09f125),
a modification was made to a xpath to include the closing attribute for the
Belgian tax report. Prior to this change, the xpath was malfunctioning but
didn't result in a traceback.

The purpose of the xpath was to retrieve the last div of the pdf_export_filters
template to ensure that the closing attribute would consistently be placed at
the end of the export filters.

However, instead of targeting the last div containing the class row as intended,
the xpath was returning all the last div HTML tags with the class col, resulting
in an output like this:
>>> [<Element div at 0x7fb94880d040>, <Element div at 0x7fb94880d000>, <Element div at 0x7fb94880d080>, <Element div at 0x7fb94880d0c0>, <Element div at 0x7fb94880d100>, <Element div at 0x7fb94880d180>, <Element div at 0x7fb94877ad00>]

By introducing parentheses, the xpath's scope was adjusted to select the last
element containing the class row, as initially intended.
Result after the fix:
>>> [<Element div at 0x7fb94877ad00>]

no task id

closes odoo/enterprise#59453

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>