[FIX] l10n_be_reports: Partner VAT listing
The aim of this commit is having a correct Partner
VAT listing report.
Before this fix, the report has a SQL query that uses
the load_more_limit value as the limit. The issue was
that we applied the limit on a set of 3 queries, as each
of these query computes one column, the report gave us
enough line but the last column was wrongly computed (as
we exceed the limit after the second query). This means that
we had an empty column. It was totally wrong.
In the same time, other bugs were found.
Report lines are by default ordered by using the res.partner
order (based on the complete name), causing that when we generate
the xml report, the groupby (from itertools) wrongly grouped lines
as we don't sort these lines on the groupby key (vat number).
To avoid this issue, we changed the groupby by using the one from
odoo.tools.

opw-3802689
opw-3766777
opw-3797584
opw-3791023

closes odoo/enterprise#59573

X-original-commit: 3205ef5f1474469edf5001e017b1acf7eab2fa42
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>