[FIX] whatsapp_event: do not crash if noone to contact
Global communication (before and after event) using whatsapp currently crash
if there is no-one to contact, as the composer asks for record IDs and receive
nothing.

When there is nobody to contact, just skip the sending and consider the
communication to be done.

Task-3732188 (Whatsapp Event: Avoid crashing registration process)

X-original-commit: odoo/enterprise@dc8cc3c59387b0ed3a9dd5d525df3021e717fd6f
Part-of: odoo/enterprise#59987