[FIX] whatsapp_event: do not crash if noone to contact
Global communication (before and after event) using whatsapp currently crash
if there is no-one to contact, as the composer asks for record IDs and receive
nothing.

When there is nobody to contact, just skip the sending and consider the
communication to be done.

Task-3732188 (Whatsapp Event: Avoid crashing registration process)

X-original-commit: 100fc02b0b50abaea069149f7db3fb8f9b44fda8
Part-of: odoo/enterprise#60098