[FIX] appointment{_crm}: stop flushing in tests
In the community PR we remove the flushing of tracking in the test setup
for CRM tests.

This highlights an issue where some tests expect to be able to create
an appointment event as any regular user but in actuality it is forbidden.

As the test identified that it was possible, we opt to fix the behavior
rather than the test.

task-3193021

Part-of: odoo/enterprise#60913