[FIX] pos_settle_due: settle due accounts correctly
Currently, when trying to settle due accounts though the PoS session, the due balance doesn't change on the customer form.

Steps to reproduce:
-------------------
* Install **Point of Sale** and **Accounting**
* Go to the **Point of Sale** App
* Open a shop session
* Navigate to a customer with a due ammount and select **Details**
* Select **Settle due accounts**
* Select any payment method
* Validate
* Close session
* Under **Orders** select **Customers**
* Select the customer for which you just settled the due accounts
> Observation:  The due amount hasn't changed.

Why the fix:
------------
The following commit introduced a new relational system to be able to access data within the PoS frontend in the same way as it is done within the ORM in python:
https://github.com/odoo/enterprise/commit/bb06fdead097bdec1582e3cdecb99fa63c2ff799

This commit modified the definition of `paylaterPaymentMethod`. Before the commit, it was defined as:
https://github.com/odoo/enterprise/blob/4a5bb11bd59dd76f2df957225bf25d38ad5beb0c/pos_settle_due/static/src/overrides/components/payment_screen/payment_screen.js#L23-L26
and would return `{id: 3, name: 'Customer Account', is_cash_count: false, use_payment_terminal: false, split_transactions: true, …}`

Now it is defined as
https://github.com/odoo/enterprise/blob/daf09109a209f87e4f2fbb3275f55095e48d8b4d/pos_settle_due/static/src/overrides/components/payment_screen/payment_screen.js#L23-L26
and returns `undefined`.

We also notice a change in the value of `this.pos.config.payment_method_ids`. Where it previously returned
```js
Proxy(Array) {0: 1, 1: 2, 2: 3}
  [[Handler]]: Object
  [[Target]]: Array(3)
    0: 1
    1: 2
    2: 3
    length: 3
  [[Prototype]]: Array(0)
  [[IsRevoked]]: false
```

It now returns
```js
Proxy(Array) {0: Base, 1: Base, 2: Base}
  [[Handler]]: Object
  [[Target]]: Array(3)
    0:  Base {models: {…}, records: Proxy(Object), model: {…}, _raw: {…}, id: 1, …}
    1: Base {models: {…}, records: Proxy(Object), model: {…}, _raw: {…}, id: 2, …}
    2: Base {models: {…}, records: Proxy(Object), model: {…}, _raw: {…}, id: 3, …}
    length: 3
  [[Prototype]]: Array(0)
  [[IsRevoked]]: false
```

Thus we need to use `.some()` to retreive the id.

opw-3802560

closes odoo/enterprise#60726

X-original-commit: b02f2e76ac7696224c0c5ce7ff6b947d7b4c9824
Related: odoo/odoo#161849
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>