[IMP] hr_payroll_account: Regroup line by rule name
Purpose
=======

The accounting entries are grouped by batch, period and rule label when you post the
journal entries from a batch (test using test_l10n_be_hr_payroll_account) to get the translations.

If there are 2 payslips, both Net Salaries are grouped under the same account.move.line, except if
the labels are not the same (eg: 1 payslip in french, 1 in dutch).

Specification
=============

By default, the rules should be grouped if it's the same code, unless an new option is enabled on
the salary rule (because we would like to keep the current behavior for deductions/reimbursement)
for instance.

closes odoo/enterprise#59513

Taskid: 3506690
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>