[FIX] web_studio: exclude x2many & binary in sortby for map & list view
Before this commit, users were able to see one2many, many2many, and binary
fields in the sortby dropdown menu in the map and list views of the studio.

Steps to reproduce:

- Project > Tasks > All Tasks.
- Click on toggle studio and navigate to View tab in the sidebar.
- In the sortby dropdown menu, all fields, including x2many and binary, were
  visible.

Observed behavior:
Users were able to select one2many, many2many, and binary fields for sorting via
the studio sidebar in the map and list views.

Expected behavior:
One should not be able to sort by one2many, many2many, and binary fields in the
map and list views via the studio.

After this commit, users will not be able to see one2many, many2many, and binary
fields in the sortby dropdown menu in the map and list views of the studio.

Task-3790886

closes odoo/enterprise#62705

X-original-commit: e1ed17bb442574a044f8280b83610ffb96bb8319
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>