[FIX] l10n_cl_edi_boletas: fix fallback status of SII result analysis
Steps to reproduce: (requires a real certificate and a real CAF)
- Install Accounting and l10n_cl_edi_boletas
- Switch to a Chilean company
- Configure the localization settings with a real certificate and a real CAF
- Create an invoice:
  * Customer: [a Chilean customer] (e.g. Blanco Martin & Asociados)
  * Document Type: (39) Boleta Electrónica
  * Product: [any]
- Confirm the invoice
- Click on "Send Now to SII"
- "SII DTE status" will be "Ask For Status" and "Send Now to SII" will change
to "Verify on SII"
- As soon as it changes to "Verify on SII", click on it
- Repeat clicking on "Verify on SII" as soon as it becomes clickable

Issue:
At some point, "SII DTE status" will switch from "Ask For Status" to "Rejected",
even if the invoice has not been officially rejected.
And as "Verify on SII" link will not be available anymore, it will not be possible
to correct the SII DTE status.

Cause:
While checking the SII result after clicking on "Verify on SII", we could receive
a response with "estado: SOK" and "estadistica: []".
This combination doesn't allow to determine a specific status for the SII DTE
and therefore falls back on the default status that is "Rejected".

Solution:
If the SII DTE status cannot be determined, we should use "Ask For Status" as
a fallback status to allow verifying it again.

opw-3781373

closes odoo/enterprise#60361

X-original-commit: 2936922950a85c61df7fcfd655bb884d4f8700f5
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>